powersim-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.26.1
scipy==1.11.3
pyyaml==6.0.1
python-dotenv==1.0.0
httpx==0.25.1
